set cpu affinity for tx/rx threads, moved tx rx to separate modules
