+++ Cargo.toml
@@ -17,6 +17,7 @@ log = "0.4"
 errno = "0.2.7"
 crossbeam-channel = "0.5.0"
 etherparse = "0.9.0"
+core_affinity = "0.5.10"
 
 [dev-dependencies]
 anyhow = "1.0.32"
@@ -30,8 +31,3 @@ serial_test = "*"
 rusty-fork = "0.3.0"
 veth-util-rs = {git = "https://github.com/seeyarh/veth-util-rs.git", branch = "main"}
 byteorder = "1"
-
-
-
-[profile.release]
-debug = true
